Contact Form (PHP)
According to pcMag a contact form is a page on a website that allows users to communicate with the site owner. The page has fields for filling in name, 
address and type of comment. On most company websites, 
email and mailing addresses are also included; however, 
the contact form provides an immediate, convenient way for users to ask the company questions.
Contact form is an essential tool in converting a user into a customer. 
Am available for collaborations!